separating expansion panel and resolution into separate components
