+++ components/templates/index.ts
@@ -7,3 +7,5 @@ export * from "./TestExpanded"
 export * from "./Pagination"
 export * from "./showStatusIcon"
 export * from "./showTestStats"
+export * from "./TestExpansionPanel"
+export * from "./TestResolution"
